fix: seperated files in utils directory
