+++ packages/salesforcedx-vscode-core/src/commands/templates/forceApexClassCreate.ts
@@ -17,7 +17,7 @@ import {
   SfdxCommandlet,
   SfdxWorkspaceChecker
 } from '../util';
-import { OverwriteComponentPrompt } from '../util/postconditionCheckers';
+import { OverwriteComponentPrompt } from '../util/overwriteComponentPrompt';
 import { LibraryForceApexClassCreateExecutor } from './executors/LibraryForceApexClassCreateExecutor';
 import {
   APEX_CLASS_DIRECTORY,
