+++ packages/salesforcedx-vscode-core/src/commands/templates/forceLightningLwcCreate.ts
@@ -21,7 +21,7 @@ import {
   SfdxWorkspaceChecker
 } from '../util';
 import { MetadataTypeGatherer } from '../util';
-import { OverwriteComponentPrompt } from '../util/postconditionCheckers';
+import { OverwriteComponentPrompt } from '../util/overwriteComponentPrompt';
 import {
   FileInternalPathGatherer,
   InternalDevWorkspaceChecker
