+++ packages/salesforcedx-vscode-core/src/commands/templates/forceVisualforceComponentCreate.ts
@@ -22,7 +22,7 @@ import {
   SfdxCommandlet,
   SfdxWorkspaceChecker
 } from '../util';
-import { OverwriteComponentPrompt } from '../util/postconditionCheckers';
+import { OverwriteComponentPrompt } from '../util/overwriteComponentPrompt';
 import { LibraryBaseTemplateCommand } from './libraryBaseTemplateCommand';
 import {
   VISUALFORCE_COMPONENT_DIRECTORY,
