+++ packages/salesforcedx-vscode-core/src/commands/templates/forceVisualforcePageCreate.ts
@@ -19,7 +19,7 @@ import {
   SfdxCommandlet,
   SfdxWorkspaceChecker
 } from '../util';
-import { OverwriteComponentPrompt } from '../util/postconditionCheckers';
+import { OverwriteComponentPrompt } from '../util/overwriteComponentPrompt';
 import { LibraryBaseTemplateCommand } from './libraryBaseTemplateCommand';
 import {
   VISUALFORCE_PAGE_DIRECTORY,
