+++ packages/salesforcedx-vscode-core/test/vscode-integration/commands/util/postconditionCheckers.test.ts
@@ -12,7 +12,7 @@ import {
 } from '@salesforce/salesforcedx-utils-vscode';
 import { expect } from 'chai';
 import * as fs from 'fs';
-import { join, normalize } from 'path';
+import { join } from 'path';
 import { createSandbox, SinonSandbox, SinonStub } from 'sinon';
 import { channelService } from '../../../../src/channels';
 import {
@@ -23,10 +23,8 @@ import {
   PathStrategyFactory,
   SfdxCommandlet
 } from '../../../../src/commands/util';
-import {
-  CompositePostconditionChecker,
-  TimestampConflictChecker
-} from '../../../../src/commands/util/postconditionCheckers';
+import { CompositePostconditionChecker } from '../../../../src/commands/util/compositePostconditionChecker';
+import { TimestampConflictChecker } from '../../../../src/commands/util/timestampConflictChecker';
 import { conflictView, DirectoryDiffResults } from '../../../../src/conflict';
 import { TimestampFileProperties } from '../../../../src/conflict/directoryDiffer';
 import { WorkspaceContext } from '../../../../src/context';
@@ -34,9 +32,9 @@ import * as workspaceUtil from '../../../../src/context/workspaceOrgType';
 import { nls } from '../../../../src/messages';
 import { notificationService } from '../../../../src/notifications';
 import { sfdxCoreSettings } from '../../../../src/settings';
-import { SfdxPackageDirectories } from '../../../../src/sfdxProject';
 import { MetadataDictionary, workspaceUtils } from '../../../../src/util';
 import { OrgType } from './../../../../src/context/workspaceOrgType';
+
 describe('Postcondition Checkers', () => {
   let env: SinonSandbox;
   describe('EmptyPostconditionChecker', () => {
